requirement.txt
web3>=6.0.0
py-solc-x>=2.0.0
liboqs-python>=0.7.0
ecdsa>=0.18.0
matplotlib>=3.7.0
numpy>=1.24.0
tabulate>=0.9.0
cryptography>=41.0.0